+++ jabberd14/jabberd/lib/socket.cc
@@ -184,7 +184,7 @@ struct in_addr *make_addr(char const* host) {
         }
     } else {
         addr.s_addr = inet_addr(host);
-        if(addr.s_addr != -1) {
+        if(addr.s_addr != (uint32_t) -1) {
             return &addr;
         }
         hp = gethostbyname(host);
